+++ name_generator/name_generator_test.go
@@ -27,3 +27,37 @@ func TestRandomNameGenerator_GenerateName(t *testing.T) {
 
 	require.Contains(t, potentialCandidates, actual)
 }
+
+// check for duplicates for adjectives
+func Test_noDuplicatesInAdjectives(t *testing.T) {
+	var duplicateArr []string
+	adjectivesHash := map[string]int{}
+
+	for _, adj := range NOUNS {
+		freq, found := adjectivesHash[adj]
+		if found && freq == 1 {
+			duplicateArr = append(duplicateArr, adj)
+		}
+
+		adjectivesHash[adj] = adjectivesHash[adj] + 1
+	}
+
+	require.Len(t, duplicateArr, 0, "Duplicate Error: found %v multiple times in ADJECTIVES", duplicateArr)
+}
+
+// check for duplicates for nouns
+func Test_noDuplicatesInNouns(t *testing.T) {
+	var duplicateArr []string
+	nounsHash := map[string]int{}
+
+	for _, noun := range NOUNS {
+		freq, found := nounsHash[noun]
+		if found && freq == 1 {
+			duplicateArr = append(duplicateArr, noun)
+		}
+
+		nounsHash[noun] = nounsHash[noun] + 1
+	}
+
+	require.Len(t, duplicateArr, 0, "Duplicate Error: found %v multiple times in NOUNS", duplicateArr)
+}
